time64.[ch]: Inline only use of a macro
This macro is now only used once.  It is accessible outside perl because
it is in time64.h, but isn't used in cpan.  Remove it and place its
expansion in time64.c.

Part of the expansion called another function.  That function can be
removed and the expansion simplified, since localtime() is now
automatically converted into localtime_r() if available and needed.