Debugging GH #17671.
We are seeing an occasional smoke failure in the experimental regular
expression Unicode property wildcard feature for the name property.
This is not reproducible, and neither valgrind nor asan shows anything.
All the failures I have seen involve looking up NULL.  Now it could be
something special about that name, or merely that it is the first in the
test to be checked.  This commit seeks to find out by skipping that
test.  If we stop getting these intermittent failures, it points one
direction; if they continue with 0x01, instead, it points in another.