t/TEST: Unset PERL_UNICODE
This commit adds this environment variable to the list of the ones that
are generally unset during the test suite execution.  Tests shouldn't
have to consider, for example, if PERL5OPT has something set like -w,
and the same for PERL_UNICODE, as its being set causes all filehandles
to have :utf8 set by default.

This should fix test suite smoke faiures we are getting in IO::Socket