regcomp.c: Rmv C undefined behavior
One analyzer said that what this commit changes was C undefined
behavior, in casting void* pointers.  Right now, the only actual type it
is called with is SV*, but I made it void*, because I thought it might
be used more generally.  But, it turns out that Unicode is planning on
changing its regular expression processing requirements to where what I
have no longer will make sense.  And, since only SV* is actually used,
this commit changes the void* to SV*, removing any undefined behavior,
with no changes to program logic.

The changes for the new Unicode direction will come in probably 5.34;
their document is still in draft, but I anticipate it will soon be
finalized.