regcomp.c: Move test earlier in regpiece()
Prior to this commit, the code checked if this was a particular
quantifier; if so handled that, then checked if it were some other
quantifier, returning if not.

This commit changes to check first if it is any quantifier, using the
macro that checks for all of them, returning if not.

Thus after the new code we know it is a quantifier and this allows later
commits to simplify