re_intuit_string() increment ref cnt of returned SV
This internal function returns an SV that it turns out may have its
reference count decremented by the next call to re_intuit_start().
Thus, the caller doesn't get clear title to the returned SV.  This was
not documented.

But it is a simple matter to increment the ref count of the returned
scalar, and document that the caller needs to decrement it.

This fixes #17734.