Refactor rninstr()
This function finds a needle in a haystack.

If memrchr() is available, use it to find the final byte in the needle;
If the needle is length 1, that's all we need, and this is special
cased.  Otherwise it uses memEQ of the other bytes of the needle, from
where the needle would have to start, and if no match, repeats.

If no memrchr(), it emulates that by just looking backward from the end
until it finds the needle's final byte; then proceeds as above.