regexec.c: Slight speed up of intuit matching
For an operation where we know there is an initial substring, the code
now looks for that whole substring before returning a candidate
position.  Previously it looked for just the first byte, and then the
condition had to include looking for the remainder.  Effectively
memmem() is now called, instead of memchr() followed by a comparison of
the rest.