regex: Add POSIXA1R node
Several of the POSIXA classes are a single range on ASCII platforms, and
[:digit:] is a single range on both ASCII and EBCDIC.  This regnode was
designed to replace the POSIXA regnode for such classes to get a bit of
performance by not needing to do an array lookup.  Instead it encodes
some bits in the flags field that with shifting and masking get the
right values for the single range's bounds for any such node.

However, performance tests conducted by Sergey Aleynikov showed this was
actually slower than what it intended to replace.  Rather than
completely drop this work, I'm adding it to blead, and immediately
reverting it, so that should parts of it ever become useful, it would be
available.

A few tests fail; those are skipped for the purposes of this commit so
that it doesn't interfere with bisecting.

The code also isn't completely commented.

One could add a regnode for each posix class it was decided should have
the expected performance boost.  But regnodes are a finite resource, and
the boost is probably not large enough to justify doing so.