regcomp,regexec: Shorten #define name
This names a flag bit whose meaning depends on context.  The previous
name contained both meanings, and was simply too long to be
comprehensible.  This commit splits it into two names, with the suffix
'_shared'.  It's not necessary to know precisely what the other meaning
is when reading the code, only that it has another meaning.