Provide a better recommendation for 'Run only under a shell'
Using an unset variable hides the true intent and
also requires an extra backslash `\$running_under_some_shell`
when used in heredoc.

Note that this could also lead to mistake when using
`\$` in a regular Perl program, as this would be true
and not false as it should be.

Stop recommending the use of an undefined variable for
the shell fallback. Use '0', with a comment
making clear the goal of 'if 0'.