Fix uninitialized fold table for EBCDIC with no locales
The C locale table for EBCDIC platforms was not getting initialized at
compile time, but only not for EBCDIC platforms.  It normally gets
changed when a new locale is switched to, but if that doesn't happen, it
doesn't get initialized at all.  IBM on z/OS has not implemented locales
with threads, so on threaded perls, the table was never getting
initialized.