Get t/porting/test_bootstrap.t to pass
Note:  It was failing because it detected two instances of 'use ' in
t/comp/redef.t.  While the insertion of 'use p5; use warnings;' might be
necessary here, it goes against the principle (which test_bootstrap.t
was exercising) of not using 'use ' statements in t/comp/*.t -- the
oldest and most basic tests of perl's functionality.