Get porting/manifest.t and porting/pod_rules.t passing
The documentation for t/porting/pod_rules.t suggests that if
t/porting/manifest.t is passing, pod_rules.t should pass as well.

However, with two files added for the MANIFEST for the core-p7 branch,
this proved not to be the case.  It turned out that we needed to add a
special hard-coded entry for README.md at one location and to modify the
code to treat two files as identical if they matched except for a
trailing newline.