Carp_overloadless.t: Restore 'use warnings'
Per discussion with Nicolas R, for distributions in the dual-life directories (dist/
and cpan/) we have to account for the possibility that the distro will be
released separately on CPAN and will need explicit 'use warnings' and/or 'use
strict' statements to continue to work as previously with Perl 5.

In this case, we not only restore 'use warnings' but have to add a 'use
strict' as well so that on Perl 5 the subsequent "no strict 'refs'" makes
sense.  (The "no warnings 'once'" was already commented-out in blead.)