ext/File-Find: improve compliance with strict
With strict-by-default (in this case, 'vars'), "import File::Find" is
read with 'import' as a global variable lacking a package name.  This
can be avoided by calling it as a method.

Suppress possible uninitialized value warning in taint.t as it reads your
filesystem.

For:  https://github.com/atoomic/perl/issues/64