Rewrite delimcpy to use memchr and Copy, not per-byte
Prior to this commit delimcpy() parsed its input byte-by-byte, looking
for a particular character, and copies the input to the output stopping
just before the first such occurrence.

memchr() is much faster for finding a single character.

The complication is that if the character is preceded by a backslash,
it doesn't count as that character, it is considered to be escaped, and
parsing continues to the first unescaped occurrence, if any.  Each
escaping backslash is not copied.

The new routine looks for the character with memchr, sees if it is
escaped.  If not, Copy does the whole copy at once.  If it is escaped,
it uses Copy up to that backslash, and repeats the process.

I suspect that it is a bug that the routine did not consider the
possibility of the escape itself being escaped, so not counting as
escaping the searched for character.  But I have retained that behavior.