lib/vars.t: Add descriptions to all tests
Mainly so that if we have to work on this file in the future, we know
where we are in the output.  "ok 17" is not sufficiently descriptive.

The file was apparently written back before test harnesses became the
norm, so we're avoiding t/test.pl and not placing repeated code into
subroutines as we normally would.