autodoc: Add ability to list variants like perlfunc does
This adds apidoc_item so that you can specify things that are closely
related in a single entry, with the list of the things at the top, much
as perlfunc does with, say, various forms of split

The arguments to the non-main entry are accepted but nothing is done
with them now.  This is because I was uncertain how we might want to
display things, as the arguments to the main entry have been shown
forever only in the usage paragraph at the end, unlike perlfunc where
the arguments are displayed on the same line as each variant.