Catch errors earlier in Devel-Peek/t/Peek.t
When testing Devel-Peek we can run some tests
inside an eval and do not bother to check if it
succeeds or not.

As STDERR is redirected to a file, we do not even
notice the error message when an error occurs.

This commit postpones the redirection of STDERR
and 'die' if the eval quoted string from `do_test` fails.