autodoc.pl: Add ability to link to other pods
This enhances perlapi and perlintern so that an API element that is
documented in some other pod can automatically have a link to that pod
generated and placed into perlapi/perlintern.  This allows one stop
browsing through the API, as the hither-to-unlisted elements now are
listed there, with a link.