Skip t/porting/bench.t if the cachegrind test takes too long.
On this armv7l system (current Raspberry Pi OS) /usr/bin/valgrind exists,
but if I try to run it it gets stuck in a CPU busy loop somewhere. As was,
this meant that t/porting/bench.t would never complete, causing the test
suite to hang.

As we are already assuming ulimit exists to suppress coredumps, also use it
to set a CPU time limit. This way, the sanity check will time out (and fail)
causing t/porting/bench.t to skip, and the regression tests to pass.

On other systems the generous timeout is harmless, and the test still runs as
intended.